Introduce Docker workflow with VS Code
Currently contributors to this project have to jump through a lot of
hoops to setup the project to run/develop it locally. To help make that
easier, introduce a VS Code Docker workflow.

All users need is Docker installed on their system to get started, and
the project builds from there.